requirements.txt
fastapi~=0.104.0
email-validator~=2.1.0
pydantic-settings~=2.1.0
pydantic-extra-types~=2.1.0
python-multipart~=0.0.6
uvicorn[standard]~=0.24.0
sqlalchemy~=2.0.0
asyncpg~=0.29.0
alembic~=1.13.0
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.0